requirements.txt
amqp==2.6.0
arrow==1.1.0
asgiref==3.3.4
astroid==2.5
bcrypt==3.1.7
beautifulsoup4==4.9.3
billiard==3.6.4.0
binaryornot==0.4.4
blinker==1.4
Brotli==1.0.9
celery==4.4.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cookiecutter==1.7.3
cryptography==3.4.7
csscompressor==0.9.5
Django==3.2.6
django-advanced-filters==1.1.1
django-appconf==1.0.4
django-bootstrap4==2.0.0
django-braces==1.14.0
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-classy-tags==2.0.0
django-compressor==2.4
django-cookie-law==2.0.4
django-cors-headers==3.2.1
django-countries==7.2.1
django-crispy-forms==1.11.2
django-crum==0.7.9
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-mathfilters==1.0.0
django-modeltranslation==0.15
django-resized==0.3.11
django-tailwind==2.2.0
django-timezone-field==4.0
django-user-visit==0.4.1
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.7.0
dnspython==1.16.0
entrypoints==0.3
environ==1.0
et-xmlfile==1.0.1
eventlet==0.25.2
factory-boy==2.12.0
Faker==4.1.0
future==0.18.2
greenlet==1.1.0
gunicorn==20.0.4
idna==2.10
isort==4.3.21
Jinja2==3.0.1
jinja2-time==0.2.0
jsonfield==3.1.0
keyring==18.0.1
kombu==4.6.11
lazy-object-proxy==1.6.0
lml==0.1.0
Markdown==3.2.2
MarkupPy==1.14
MarkupSafe==2.0.1
mccabe==0.6.1
monotonic==1.6
numpy==1.20.3
opencv-python==4.4.0.46
openpyxl==3.0.7
pandas==1.1.0
Pillow==8.2.0
poyo==0.5.0
psycopg2-binary==2.8.6
pycparser==2.20
pyexcel-io==0.5.20
pyexcel-xlsx==0.5.8
PyJWT==2.1.0
pylint==2.5.3
pytest-runner==5.3.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-slugify==5.0.2
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
rabbitmq==0.2.0
rcssmin==1.0.6
reportlab==3.5.67
requests==2.25.1
retrying==1.3.3
rjsmin==1.1.0
schedule==0.6.0
SecretStorage==2.3.1
simplejson==3.17.2
six==1.14.0
soupsieve==2.2.1
sqlparse==0.4.1
stripe==2.60.0
text-unidecode==1.3
toml==0.10.2
ua-parser==0.10.0
urllib3==1.26.4
user-agents==2.2.0
vine==1.3.0
whitenoise==5.0.1
wrapt==1.12.1
